requirements.txt
pandas
konlpy
gradio
pyyaml
openpyxl